Add OverlappingFluidSolidMap
This class encapsulates building up the mapping between
overlapping fluid and solid elements. This is specifically intended
for ImmersedBoundary. This can probably be generalized but starting
with something specific to get going. Note that this is only applicable
for ReplicatedMesh - this will break with DistributedMesh since we
loop over all elements; we throw and error if(!mesh.is_serial()).